added distance functions to sort airports

+++ voyager-migration/src/main/java/org/voyager/service/AirportService.java
@@ -7,6 +7,6 @@ public interface AirportService<T> {
     public List<String> getAllIataCodes();
     public List<String[]> getAlliataCodesAndNames();
     public List<T> getByCountryCode(String countryCode, int limit);
-    public List<T> getClosest(float latitude, float longitude, int limit);
+    public List<T> getClosest(double latitude, double longitude, int limit);
     public T getByIATA(String iata);
 }
